+++ transaction/merklepath_test.go
@@ -96,12 +96,12 @@ func TestMerklePathComputeRootHex(t *testing.T) {
 type MyChainTracker struct{}
 
 // Implement the IsValidRootForHeight method on MyChainTracker.
-func (mct MyChainTracker) IsValidRootForHeight(root *chainhash.Hash, height uint32) bool {
+func (mct MyChainTracker) IsValidRootForHeight(root *chainhash.Hash, height uint32) (bool, error) {
 	// Convert BRC74Root hex string to a byte slice for comparison
 	// expectedRoot, _ := hex.DecodeString(BRC74Root)
 
 	// Assuming BRC74JSON.BlockHeight is of type uint64, and needs to be cast to uint64
-	return root.String() == BRC74Root && height == BRC74JSON.BlockHeight
+	return root.String() == BRC74Root && height == BRC74JSON.BlockHeight, nil
 }
 
 func TestMerklePath_Verify(t *testing.T) {